% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_limma}
\alias{DEG_limma}
\title{Differentially expressed genes analysis through the limma package.}
\usage{
DEG_limma(dataset, design, contrast.matrix)
}
\arguments{
\item{dataset}{dataframe of expression values with samples in columns and genes in row.}

\item{design}{vector of 0 and 1 values. 0 for the first experimental condition, 1 for the second one.}

\item{contrast.matrix}{design matrix like one produced by the make_designMatrix() function.}
}
\value{
A dataframe with 3 columns is returned. It contains the Log Fold change value (logFC),
The Pvalue associated with this Pvalue (PValue)
the last column corresponds to the gene names given in the dataset (SYMBOL)
}
\description{
Uses the lm.fit(),eBayes() and topTable() from the limma package to automate the steps of the analysis.
}
\examples{
# Import a dataset
Data = Simul.data(type = "Microarrays", n.cond1 = 15, n.cond2 = 15, nb.genes = 1000)
#Construct the design matrix
design = make_designMatrix(dataset = Data,ncond1 = 15, ncond2 = 15)
res.DEG = DEG_limma(Data,design)
}
