% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebayesGEO.R
\name{eBayesGEO}
\alias{eBayesGEO}
\title{ebayesGEO}
\usage{
eBayesGEO(
  fit,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  proportion_vector = 0.01
)
}
\description{
Functions from supplementary files of Lu et al.2016. For more informations see : 
Lu L., Townsend K.A., et Daigle B.J.GEOlimma  :  differential  expression  
analysis and feature selection using pre-existing microarray data.BMC 
Bioinformatics 22, n 1(3 f ́evrier 2021) : 44.
https ://doi.org/10.1186/s12859-020-03932-5.
}
