% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDEG.R
\name{Get.DEG.2}
\alias{Get.DEG.2}
\title{Extract DEG from a binary matrix.}
\usage{
Get.DEG.2(binary.matrix, method, alternative)
}
\arguments{
\item{binary.matrix}{Dataframe or matrix containing only 0 and 1 values, with genes in rows and methods in columns}

\item{method}{Character string 
"union" or "intersect", "union" to get all the predicted DEG and "intersect" for the intersection of predicted DEG among the set of methods}

\item{alternative}{Logical value. 
TRUE to compute the union or intersection of predicted DEG if the binary matrix is composed of strong hypothesis prediction (for Upregulated and Downregulated genes)
FALSE if the dataframe contains no such information, just weak hypothesis prediction}
}
\value{
A list of two vectors of character string is returned for alternative = TRUE : for upregulated and downregulated genes. Otherwise, only a vector of DEG is returned
}
\description{
Analyze the output of Upset.Binary.Dataframe() to obtain Dthe union or the intersection of predicted DEG of each method.
}
\examples{
# Getting a binary matrix, DEG are designed by "1" 
# 250 genes for 10 methods (up and downregulated)
Data = matrix(runif(5000, 0, 1), ncol=20)
Binary.matrix = round(Data)
methods = paste0(rep("method", each = 20),rep(c(1:10),each = 2),rep(c("_Up","_Down"),each=1))
genes = paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Binary.matrix) = methods
row.names(Binary.matrix) = genes
# Retrieving all the predicted down and upregulated genes  
DEG = Get.DEG.2(Binary.matrix,alternative = TRUE, method = "union")
}
