% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDEG.R
\name{Get.DEG}
\alias{Get.DEG}
\title{Extract DEG from a binary matrix.}
\usage{
Get.DEG(binary.matrix, method, alternative)
}
\arguments{
\item{binary.matrix}{Dataframe or matrix containing only 0 and 1 values}

\item{method}{Character string that corresponds to any columns of the binary matrix.}

\item{alternative}{Logical value. 
If TRUE, Alternative hypothesis will be searched : for both up and downregulated. 
In this case, the method argument should be the name of the analysis (DESeq2 to extract up and downregulated genes).
If FALSE, only the method given in argument is analyzed.}
}
\value{
A list of one vector if alternative = FALSE, 2 vectors if alternative = TRUE
}
\description{
Analyze the output of Upset.Binary.Dataframe() to obtain DEG for one method
}
\examples{
# Getting a binary matrix, DEG are designed by "1" 
# 250 genes for 10 methods (up and downregulated)
Data = matrix(runif(5000, 0, 1), ncol=20)
Binary.matrix = round(Data)
methods = paste0(rep("method", each = 20),rep(c(1:10),each = 2),rep(c("_Up","_Down"),each=1))
genes = paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Binary.matrix) = methods
row.names(Binary.matrix) = genes

# Retriving both up and down regulated genes for the "method5"
DEG = Get.DEG(Binary.matrix,alternative = TRUE, method = "method5")
}
