% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoExpression.R
\name{Make.adjacency.table}
\alias{Make.adjacency.table}
\title{Computing an adjacency table}
\usage{
Make.adjacency.table(data, method)
}
\arguments{
\item{data}{Dataset of gene expression levels with genes in row and samples in columns}

\item{method}{Methods to compute the correlation coefficients.
"spearman" compute the Spearman's rho, 
"kendall" uses the Kendall's tau and
"pearson" the Pearson's product moment correlation coefficient.
These functions are called via the cor() function in the stats package.
"TOM" uses the TOMsimilarityFromExpr() function from the WGCNA package.}
}
\value{
Datraframe of three columns. Two first contains pair of genes, and the third one the correlation coefficient or the TOM similarity.
}
\description{
This function is similar to Cor.square.matrix() and TOM.square.matrix() as it converts the square matrix into a 
3 columns dataframe with no repeated pair of variables.
}
\examples{
# Creating a dataset
df = matrix(runif(500, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=1))
colnames(df) = group
row.names(df) = genes

# computing correlation for pairs of genes with Spearman's rho
Adj = Make.adjacency.table(df,method = "spearman")
}
