% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PCA_tools}
\alias{PCA_tools}
\title{Compute a PCA of the individuals of a dataframe.}
\usage{
PCA_tools(data.to.comp)
}
\arguments{
\item{data.to.comp}{dataframe containing pvalues of genes being differentially expressed, 
with tools used in row, and genes in columns}
}
\value{
Principal component analysis plot with cos2 values for each methods
}
\description{
Compute a PCA of the different tool used in normalization and DEG analysis step to discriminate each methods by their found p-values on each genes 
The function calls \link[FactoMineR]{PCA} then \link[factoextra]{fviz_pca}.
}
\examples{
# Import the Pvalues dataframe for each method
Pvalues = matrix(runif(5000, 10, 100), ncol=10)

method = paste0(rep(c("method"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=20), rep(c(1:10),each = 1))
colnames(Pvalues) = method
row.names(Pvalues) = genes 

# Plotting PCA on methods
PCA_tools(t(Pvalues))
}
