% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{relations.comparison}
\alias{relations.comparison}
\title{Producing an igraph network with different colors to compare two networks with igraph}
\usage{
relations.comparison(g1, g2, g1.name, g2.name, diplay)
}
\arguments{
\item{g1}{igraph class object.}

\item{g2}{igraph class object.}

\item{g1.name}{Character string to be displayed on the legend of the graph for the first graph.}

\item{g2.name}{Character string to be displayed on the legend of the graph for the second graph.}

\item{diplay}{Logical value.
TRUE if you want to to diplay the graph, otherwise it will be stocked as an object.}
}
\value{
igraph class object with colored edges.
}
\description{
For two graphs, red edges will be the common ones. It is used to compare two methods with one another, or two experimental conditions of a same set of genes for example.
}
\examples{
# Creating two datasets
df = matrix(runif(500, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=1))
colnames(df) = group
row.names(df) = genes
# Computing relation network
G.spearman = Make.df.graph(df1, cor.threshold = 0.5, Pvalue.threshold = F, method = "spearman")
G.TOM = Make.df.graph(df1, cor.threshold = 0.05, Pvalue.threshold = F, method = "TOM")
# Comparing both relation network
G.comp = relations.comparison(g1 = G.spearman, g2 = G.TOM, g1.name = "Spearman", g2.name = "TOM similarity", diplay = T)
}
