% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoExpression.R
\name{TOM.square.matrix}
\alias{TOM.square.matrix}
\title{Computing a square matrix of Topological Overlap}
\usage{
TOM.square.matrix(data)
}
\arguments{
\item{data}{Dataset of gene expression levels with genes in rows and samples in columns}
}
\value{
square matrix of TOM similarity measurements for each genes
}
\description{
Uses the \link[WGCNA]{TOMsimilarityFromExpr} function from the WGCNA package
}
\examples{
# Creating a dataset
df = matrix(runif(500, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=1))
colnames(df) = group
row.names(df) = genes

# Computing TOM similarity
TOM = TOM.square.matrix(df)
}
