% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tools.DEG.Microarray.merge}
\alias{tools.DEG.Microarray.merge}
\title{Merge normalization step with statistical methods to analyse DE genes}
\usage{
tools.DEG.Microarray.merge(dataset, design, tools.norm, tools.DEG)
}
\arguments{
\item{dataset}{AffyBatch or GEO id (e.g. "GSE31684" )}

\item{design}{design matrix produce by the function \link[stats]{model.matrix}}

\item{tools.norm}{Character string among : 
"rma", "gcrma", "mas5","none", "liwong", "RMA.inv.mas", "mas.mas.inv.mas", 
"mas.mas.const.liwong", "mas.mas.inv.med", "mas.mas.inv.liwong"
"rma" calls \link[affy]{rma}.
"mas5" calls \link[affy]{mas5}.
"gcrma" calls \link[gcrma]{gcrma}.
The other methods calls \link[affy]{expresso}.
"liwong" has the following parameters : (bgcorrect.method= "none", 
normalize.method= "invariantset", pmcorrect.method= "pmonly", summary.method= "liwong")
Methods that start with "RMA." has the parameters : 
bgcorrect.method = "rma" , pmcorrect.method = "pmonly" 
Those that starts with "mas.mas" have the following parameters : 
bgcorrect.method = "rma" , pmcorrect.method = "pmonly" 
".inv" is for  normalize.method = "invariantset", 
".const" is for normalize.method = "constant" 
".liwong" is for summary.method= "liwong"
".med" is for summary.method= "medianpolish"
".mas" is for summary.method= "mas"}

\item{tools.DEG}{Different methods used to compute pvalues of differentially expressed genes for microarrays
"Wilcox" uses the \link[GENEXPRESSO]{wilcoxDEG} function 
"limma" and "GEOlimma" calls \link[limma]{eBayes} and \link[limma]{lmFit} 
"RankProduct" calls \link[RankProd]{RankProducts} with calculateProduct = TRUE,
"RankSum" calls \link[RankProd]{RankProducts} with calculateProduct = FALSE}
}
\value{
a dataframe with pvalues of genes being DE depending on the used methods
}
\description{
Merge normalization step with statistical methods to analyse DE genes
}
\examples{
#celpath = file.path("Insert/your/path/to/.CEL/files")
#txt.dir = paste0(celpath,"/tab.txt")
#tab = read.delim(txt.dir,check.names=FALSE,as.is=TRUE, header = T, fill = TRUE)

# Here to compare T1 vs T2 patients
#tab = subset(tab, tab$PreOpClinStage == 'T1' | tab$PreOpClinStage == 'T2') 
#creating the model matrix
#design = model.matrix(~0 + tab$PreOpClinStage)

# computing pvalues for DE genes with one normalization in combination with 
# 3 different statistical analysis
#DEG = tools.DEG.Microarray.merge(dataset = abatch, 
#                                 design = design, 
#                                 tools.norm = c("rma"), 
#                                 tools.DEG = c("GEOlimma","limma","RankProduct") )
}
