% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tools.DEG.Microarray}
\alias{tools.DEG.Microarray}
\title{Compute the pvalues of genes being up- and down-regulated for Microarray datasets.}
\usage{
tools.DEG.Microarray(data, tool, n1, n2, design)
}
\arguments{
\item{data}{Dataframe of gene expression levels with sample names in columns and genes in rows}

\item{tool}{Different methods used to compute pvalues of differentially expressed genes for microarrays
"Wilcox" uses the \link[GENEXPRESSO]{WilcoxDEG} function implemented in this very same pacakge 
"limma" and "GEOlimma uses respectively the functions \link[GENEXPRESSO]{DEG_limma} and \link[GENEXPRESSO]{DEG_GEOlimma} that comes from the limma pacakge
"RankProduct","RankProduct.log" perform a Rank Product analysis with the \link[RankProd]{RankProducts} function from the RankProd package for normal and logged values respectively 
"RankSum","RankSum.log" perform a Rank Sum analysis with the \link[RankProd]{RankProducts} function from the RankProd package for normal and logged values respectively}

\item{n1}{Number of samples for the first experimental condition. (Optionnal)}

\item{n2}{Number of samples for the second experimental condition. (Optionnal)}
}
\value{
A dataframe with genes in rows and pvalues of a gene being upregulated and downregulated in columns
}
\description{
Through different functions contained in several packages, this function computes pvalues of differentially expressed genes
}
\examples{
# Import the dataset
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 

# Creating a model matrix :
status = c(rep("control",10),rep("case",10))
design = model.matrix(~0+status) 


# Compute Pvalues for the RankProduct method
res.DEG = tools.DEG.Microarray(Data,"RankProduct",design = design)
}
