% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tools.DEG.Nanostring}
\alias{tools.DEG.Nanostring}
\title{Compute the pvalues of genes being up and downregulated for Nanostring datasets.}
\usage{
tools.DEG.Nanostring(raw.data, tool, data, tool_norm)
}
\arguments{
\item{raw.data}{rcc type file. List of 4 elements: 
Samples.IDs is a dataframe with four columns: sample name, status (WildType, Mutated), filename. 
rcc.df contains expression levels with samples in column and genes in rows. 
annots.df is a dataframe with 3 columns: geneclass (endogenous, housekeeping, negative, positive), gene name, and the accession number. 
FOV is a dataframe that contains Field of views information.}

\item{tool}{Method to use to compute the pvalues of differentially expressed genes.
"Wilcox" uses the wilcoxDEG() function implemented in this very same pacakge 
"limma" uses the functions DEG_limma() that comes from the limma pacakge
"RankProduct" and "RankSum" perform respectively a Rank Product and a Rank Sum analysiswith the RankProducts() function from the RankProd package
"desq2" uses the DESeq() function from the DESeq2 package. This last one is particular, because if it is chosen it will ignore the argument "tool_norm"}

\item{data}{Dataframe of gene expression levels with sample names in columns and genes in rows}

\item{tool_norm}{Normalization tool used previously in the tools.norm.Nanostring() function}
}
\value{
A dataframe with genes in rows and pvalues of a gene being upregulated and downregulated in columns
}
\description{
Through different functions contained in several packages, this function computes pvalues of differentially expressed genes
}
