% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tools.DEG.RNAseq.merge}
\alias{tools.DEG.RNAseq.merge}
\title{Combine Normalization with DEG analysis}
\usage{
tools.DEG.RNAseq.merge(count.matrix, tools.norm, tools.DEG, design)
}
\arguments{
\item{count.matrix}{Dataframe of count with samples in columns and genes SYMBOL in rows.}

\item{tools.norm}{Character string among "TMM","TMMwsp", "RLE", "Upperquartile", "voom", "vst", "vst2".
"TMM","TMMwsp", "RLE", "Upperquartile" calls the \link[edgeR]{calcNormFactors} function.
"voom" calls the \link[limma]{voom} function.
"vst" calls the \link[DESeq]{estimateSizeFactors} function on a CountDataSet.
"vst2" does the same but also calls the \link[DESeq2]{varianceStabilizingTransformation} function.}

\item{tools.DEG}{Character string among : "ExactTest", "GLM", "nbinom", "nbinom.Wad", "nbinom.LRT"
"ExactTest calls the \link[edgeR]{exactTest} function.
"GLM" uses a linear model with the \link[edgeR]{glmQLFit} and \link[edgeR]{glmQLFTest} functions.
"nbinom" is the DESeq equivalent with the \link[DESeq]{nbinomTest} function.
"nbinom.Wald" and "nbinom.LRT" calls the same function with different parameters : \link[DESeq2]{DESeq}.}

\item{design}{Vector of 1 and 2 of the same length of colnames(count.matrix).
1 for the first group and 2 for the second.}
}
\value{

}
\description{
Calls the normalization then, a set of DEG analysis to create several combinations.
}
\examples{
# load a count matrix (example with a random dataset)
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 

# Compute design vector
design = c(rep(1,10), rep(2,10)) # 10 from group 1, 10 from group 2

# Test all combinations :
#DEG = tools.DEG.RNAseq.merge(Data,"all","all",design)

# Test only a few :
DEG = tools.DEG.RNAseq.merge(Data,c("TMM","voom","vst"),c("ExactTest","nbinom.Wald"),design)
}
