% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tools.DEG.RNAseq}
\alias{tools.DEG.RNAseq}
\title{Compute the pvalues of genes being differentially expressed for RNA-seq type data.}
\usage{
tools.DEG.RNAseq(data, tool)
}
\arguments{
\item{data}{dataframe of gene expression levels, with genes in rows and samples in columns.}

\item{tool}{Method to Normalize datasets and statistically analysing them
"edgeR_RLE","edgeR_upperquartile","edgeR_TMMwsp", and "edgeR_TMM" are methods of normalization implemented in the edgeR package with the function calcNormFactors.DGEList()
If "tool" is one of these parameters, it will compute two analysis: the exact test from the exactTest() function 
and a Quasi-likelihood test with the glmQLFTest() function still in the edgeR package.
"deseq2.Wald" and "deseq2.LRT' uses the same normalization methods contained in the DESeq2 package with de DESeq() function.
The first method uses a Wald test and the second a Likelihood ratio test to determine Pvalues of differentially expressed genes.}
}
\value{
Dataframe with genes in row, and methods used in columns. In contains the differentially expressed p-values for each gene.
}
\description{
Through different functions contained in several packages, this function computes pvalues of differentially expressed genes
}
\examples{
# Import a RNA seq dataset
Data = Simul.data(type = "RNAseq", n.cond1 = 15, n.cond2 = 15, 
                  nb.genes = 100)
# computing pvalues of DEG with TMM normalization method
res.DEG = tools.DEG.RNAseq(data = Data, tool = "edgeR_TMM")
}
