% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{tools.norm.Nanostring}
\alias{tools.norm.Nanostring}
\title{Normalize a Nanostring dataset}
\usage{
tools.norm.Nanostring(raw.data, tool, nanoR = F)
}
\arguments{
\item{raw.data}{rcc type file. List of 4 elements: 
Samples.IDs is a dataframe with four columns: sample name, status (WildType, Mutated), filename. 
rcc.df contains expression levels with samples in column and genes in rows. 
annots.df is a dataframe with 3 columns: geneclass (endogenous, housekeeping, negative, positive), gene name, and the accession number. 
FOV is a dataframe that contains Field of views information.}

\item{tool}{Normalization tool. "nappa.NS", "nappa.param1","nappa.param2","nappa.param3" are different parameters used with the NAPPA() function from the NAPPA package.
"nanostringnorm.default","nanostringnorm.param1","nanostringnorm.param2" use the NanoStringNorm() normalization function from the package NanoStringNorm
"nanostringR" uses the HKnorm() function from the package nanostringr.
"nanoR.top100","nanoR.total" uses the nsNormalize() function from the nanoR package.}

\item{nanoR}{Logical value. TRUE is necessary if the tool used is "nanoR.top100" or "nanoR.total". By default, the value is set on FALSE}
}
\value{
dataframe of normalized expression values
}
\description{
Normalization through different methods in several packages
}
\examples{

# Normalizing data using one method :
Norm.data = tools.norm.Nanostring(raw.data = data, tool = "nappa.NS")
}
