% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{wilcoxDEG}
\alias{wilcoxDEG}
\title{Multiple Wilcoxons tests for each row of a dataframe}
\usage{
wilcoxDEG(data, n1, n2)
}
\arguments{
\item{data}{dataframe of gene expression levels: Gene names in rows, samples in columns.}

\item{n1}{Number of samples for the first experimental condition}

\item{n2}{Number of samples for the second experimental condition}
}
\value{
Dataframe with three columns, each corresponding to a hypothesis : 
weak hypothesis, and two for the strong hypothesis "alternative = "less" ", and "greater"
For the second condition has values that are less or greater than values of the first one.
}
\description{
Testing, for one row at the time, if the first series of values are different, greater or less than the values of the second condition.
}
\examples{
# Import a dataset
Data = Simul.data(type = "Microarrays", 
                  n.cond1 = 15,  
                  n.cond2 = 15, 
                  nb.genes = 100)
# Compute Pvalues
res.DEG = wilcoxDEG(Data,15,15)
}
