% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_alternative}
\alias{DEG_alternative}
\title{Takes the results of DEG_GEOlimma and DEG_limma to extract the pvalues of alternative hypothesis.}
\usage{
DEG_alternative(res.diff_limma)
}
\arguments{
\item{res.diff_limma}{A dataframe with 3 columns is returned. It contains the Log Fold change value (logFC),
The Pvalue associated with this Pvalue (PValue),
the last column corresponds to the gene names (SYMBOL).}
}
\value{
A dataframe with 3 columns is returned. The two columns corresponds to the pvalue
of a gene being respectively, more expressed and less expressed in the second condition
}
\description{
Takes the results of DEG_GEOlimma and DEG_limma to extract the pvalues of alternative hypothesis.
}
\examples{
#get the results of limma analysis
res.diff = DEG_limma(dataset,design, contrast.matrix = cm)
#compute the results of alternative hypothesis
res.diff.alternative = DEG_alternative(res.diff)
}
