% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_alternative}
\alias{DEG_alternative}
\title{Takes the results of DEG_GEOlimma and DEG_limma and extracts the pvalues of alternative hypothesis.}
\usage{
DEG_alternative(res.diff_limma)
}
\arguments{
\item{res.diff_limma}{A dataframe with 3 columns is returned. It contains the Log Fold change value (logFC),
The Pvalue associated with this Pvalue (PValue),
the last column corresponds to the gene names (SYMBOL).}
}
\value{
A dataframe with 3 columns is returned. The two columns corresponds to the pvalue
of a gene being more expressed and less expressed, in the second condition
}
\description{
Takes the results of DEG_GEOlimma and DEG_limma and extracts the pvalues of alternative hypothesis.
}
\examples{
# Import the dataset
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 

#Build the design matrix
#design = make_designMatrix(dataset = Data,ncond1 = 10, ncond2 = 10)
#get the results of limma analysis
#res.diff = DEG_limma(dataset = Data,design)

#compute the results of alternative hypothesis
#res.diff.alternative = DEG_alternative(res.diff)
}
