% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebayesGEO.R
\name{eBayesGEO}
\alias{eBayesGEO}
\title{ebayesGEO}
\usage{
eBayesGEO(
  fit,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  proportion_vector = 0.01
)
}
\description{
ebayesGEO
}
