% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoExpression.R
\name{Make.full.adjacency}
\alias{Make.full.adjacency}
\title{Computing an adjacency table with or without Pvalues, with Spearman, Kendall correlation coefficient and TOM similarity.}
\usage{
Make.full.adjacency(data, PValue = T)
}
\arguments{
\item{data}{Dataset of gene expression levels with genes in row and samples in columns}

\item{PValue}{Logical value. 
if PValue = TRUE, all the pvalues associated with the corresponding coeffient of correlation will be computed. Could be very long on large dataframe
if PValue = FALSE, only the coefficient of correlation, and TOM similarity will be computed.}
}
\value{
Dataframe with several columns. the two first are the pairs of genes,
the other are the coefficient of correlation depending on the used method, and maybe the PValues depending on the argument.
}
\description{
Calling Make.adjacencyPVal() with all the implemented methods.
}
\examples{
# Creating a dataset
df = matrix(runif(500, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=1))
colnames(df) = group
row.names(df) = genes

# computing fast correlation and pvalues 
Adj = Make.full.adjacency(df,PValue = TRUE)
}
