% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{make_designMatrix}
\alias{make_designMatrix}
\title{Creates a Model matrix.}
\usage{
make_designMatrix(
  dataset,
  cond1 = "A",
  cond2 = "B",
  ncond1 = (ncol(dataset)/2),
  ncond2 = (ncol(dataset)/2)
)
}
\arguments{
\item{dataset}{dataframe of expression values with samples in columns and genes in row.}

\item{cond1}{Name of the first experimental condition.}

\item{cond2}{Name of the second experimental condition.}

\item{ncond1}{Number of sample in the first experimental condition.}

\item{ncond2}{Number of sample in the the second experimental condition.}
}
\value{
Model matrix.
}
\description{
Creates a Model matrix.
}
\examples{
# Using an existing dataset
Data = Simul.data(type = "RNAseq", n.cond1 = 15, n.cond2 = 15, nb.genes = 1000)
design = make_designMatrix(dataset = Data, cond1 = "control", cond2 = "Case", ncond1 = 15, ncond2 = 15)
}
