% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{mapping.affymetrix.probe}
\alias{mapping.affymetrix.probe}
\title{Mapping Probes ID of Affymetrix microarray chip in expression set dataframe}
\usage{
mapping.affymetrix.probe(exprSet)
}
\arguments{
\item{exprSet}{Dataframe with samples in columns, probes ID in rows}
}
\value{
Dataframe with about 20.000 rows. Gene symbols will replace the probes ID
}
\description{
This function works only for microarray chip "hgu133plus2"
}
