% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PCA_tools}
\alias{PCA_tools}
\title{Compute a PCA of the different tool used in normalization and DEG analysis step to discriminate each methods by their found p-values on each genes}
\usage{
PCA_tools(data.to.comp)
}
\arguments{
\item{data.to.comp}{datarame containing pvalues of genes being differentially expressed, 
with tools used in row, and genes in columns}
}
\value{
Principal component analysis plot with cos2 values for each methods
}
\description{
Compute a PCA of the different tool used in normalization and DEG analysis step to discriminate each methods by their found p-values on each genes
}
\examples{
# Import a dataset (Microarrays for example)
Data = Simul.data(type = "Microarrays", n.cond1 = 15, 
                                        n.cond2 = 15, 
                                        nb.genes = 100)
# Compute Pvalues for all the methods 
res.DEG = tools.DEG.Microarrays.merge(Data)
# Plotting PCA on methods
PCA = PCA_tools(res.DEG)
}
