% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PCA_tools}
\alias{PCA_tools}
\title{Compute a PCA of the different tool used in normalization and DEG analysis step to discriminate each methods by their found p-values on each genes}
\usage{
PCA_tools(data.to.comp)
}
\arguments{
\item{data.to.comp}{dataframe containing pvalues of genes being differentially expressed, 
with tools used in row, and genes in columns}
}
\value{
Principal component analysis plot with cos2 values for each methods
}
\description{
Compute a PCA of the different tool used in normalization and DEG analysis step to discriminate each methods by their found p-values on each genes
}
\examples{
# Import the dataset
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 

# Compute Pvalues for all the methods 
# res.DEG = tools.DEG.RNAseq.merge(Data)
# Plotting PCA on methods
# PCA_tools(res.DEG)
}
