% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{relations.comparison}
\alias{relations.comparison}
\title{Produce an igraph network with different colors to compare two networks with igraph}
\usage{
relations.comparison(g1, g2, g1.name, g2.name, display, color.g1, color.g2)
}
\arguments{
\item{g1}{igraph class object.}

\item{g2}{igraph class object.}

\item{g1.name}{Character string to be displayed on the legend of the graph for the first graph.}

\item{g2.name}{Character string to be displayed on the legend of the graph for the second graph.}

\item{display}{Logical value.}

\item{color.g1}{Character string for a color to be displayed onto the g1 edges.
By default the color is "blue".}

\item{color.g2}{Character string for a color to be displayed onto the g2 edges.
By default the color is "darkgreen".

TRUE if you want to to display the graph, otherwise it will be stocked as an object.}
}
\value{
igraph class object with colored edges.
}
\description{
For two graphs, red edges will be the common ones. It is used to compare two methods with one another, or two experimental conditions of a same set of genes for example.
}
\examples{
# Creating two datasets
# Import the dataset
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 
# Computing relation network
#G.spearman = Make.df.graph(Data, cor.threshold = 0.25, Pvalue.threshold = F, 
#                                                      method = "spearman")
#G.kendall = Make.df.graph(Data, cor.threshold = 0.25, Pvalue.threshold = F, 
#                                                  method = "kendall")
# Comparing both relation network
#G.comp = relations.comparison(g1 = G.spearman, g2 = G.kendall,  
#                              g1.name = "Spearman", 
#                              g2.name = "kendall", 
#                              display = T)
}
