% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Simul.data}
\alias{Simul.data}
\title{Creates a dataset}
\usage{
Simul.data(type, n.cond1, n.cond2, nb.genes)
}
\arguments{
\item{type}{Data type wanted. "RNAseq" Simulates data with SimulateRnaSeqData() from the DEFormats package. "Microarrays" simulates data with the function madsim() from madsim package. "Nanostring" imports an already existing dataset, from real data.}

\item{n.cond1}{Sample number in the first group}

\item{n.cond2}{Sample number in the second group}

\item{nb.genes}{Gene numbers}
}
\value{

}
\description{
Creates a dataset depending on the wanted data type
}
\examples{
# To get a RNAseq type data with 1000 genes and 2 groups of 15 samples
Data = Simul.data(type = "RNAseq", n.cond1 = 15, n.cond2 = 15, nb.genes = 1000)
}
