% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tools.DEG.Microarrays}
\alias{tools.DEG.Microarrays}
\title{Compute the pvalues of genes being up and downregulated for Microarray datasets.}
\usage{
tools.DEG.Microarrays(data, tool, n1, n2)
}
\arguments{
\item{data}{Dataframe of gene expression levels with sample names in columns and genes in rows}

\item{tool}{Different methods used to compute pvalues of differentially expressed genes for microarrays
"Wilcox" uses the wilcoxDEG() function implemented in this very same pacakge 
"limma" and "GEOlimma uses respectively the functions DEG_limma() and DEG_GEOlimma() that comes from the limma pacakge
"RankProduct","RankProduct.log" perform a Rank Product analysis with the RankProducts() function from the RankProd package for normal and logged values respectively 
"RankSum","RankSum.log" perform a Rank Sum analysis with the RankProducts() function from the RankProd package for normal and logged values respectively}

\item{n1}{Number of samples for the first experimental condition}

\item{n2}{Number of samples for the second experimental condition}
}
\value{
A dataframe with genes in rows and pvalues of a gene being upregulated and downregulated in columns
}
\description{
Through different functions contained in several packages, this function computes pvalues of differentially expressed genes
}
\examples{
# Import a dataset
Data = Simul.data(type = "Microarrays", 
                   n.cond1 = 15, 
                   n.cond2 = 15, 
                   nb.genes = 100)
# Compute Pvalues for the RankProduct method
res.DEG = tools.DEG.Microarrays(Data,"RankProduct",15,15)
}
