% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{tools.DEG.RNAseq.merge}
\alias{tools.DEG.RNAseq.merge}
\title{Merge the DEG Pvalues for each tool used by tools.DEG.RNAseq in one dataframe}
\usage{
tools.DEG.RNAseq.merge(data, tools)
}
\arguments{
\item{data}{Dataframe with genes in row, and methods used in columns. In contains the differentially expressed p-values for each gene.}

\item{tools}{list character string.
By default all the methods present in tools.DEG.RNAseq() are used.
Any tools given in this list could be apssed as argument : 
"edgeR_RLE","edgeR_upperquartile","edgeR_TMMwsp","deseq2.Wald","deseq2.LRT", "deseq"}
}
\value{
Dataframe of DEG pvalues with genes in columns and tools in rows.
}
\description{
Merge the DEG Pvalues for each tool used by tools.DEG.RNAseq in one dataframe
}
\examples{
# Import the dataset
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 

# Compute all the possible pvalues
# res.DEG = tools.DEG.RNAseq.merge(Data)

# Compute the pvalues, only with edgeR methods
tools = c("edgeR_RLE","edgeR_upperquartile","edgeR_TMM")
res.DEG = tools.DEG.RNAseq.merge(data = Data, tools = tools)
}
