% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{tools.norm.Microarray}
\alias{tools.norm.Microarray}
\title{Normalize a Nanostring dataset}
\usage{
tools.norm.Microarray(
  GEOiD,
  FetchOnGEOdb = FALSE,
  tools,
  tools.normalize,
  tools.bgcorrect,
  tools.pmcorrect,
  tools.express.summary.stat
)
}
\arguments{
\item{GEOiD}{GEO accession number or directory. 
If you give a GEO accession number, then set FetchOnGEOdb = TRUE.
Otherwise, if it is a directory, ignore the FetchOnGEOdb parameter. Note that if the data are stored localy, it should be inside a .tar archive.}

\item{FetchOnGEOdb}{logical value. if TRUE, GEOiD parameter should be a GEO accession number.}

\item{tools}{}

\item{tools.normalize}{}

\item{tools.bgcorrect}{}

\item{tools.pmcorrect}{}

\item{tools.express.summary.stat}{}
}
\value{

}
\description{
Normalize a Nanostring dataset
}
