% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Upset.Binary.Dataframe}
\alias{Upset.Binary.Dataframe}
\title{Compute a usable matrix to make an Upset plot}
\usage{
Upset.Binary.Dataframe(data.to.comp, threshold)
}
\arguments{
\item{data.to.comp}{Dataframe of DEG pvalues with genes in columns, and methods in rows.}

\item{threshold}{Threshold value to fill the dataframe with binary values. By default threshold = 0.05}
}
\value{
dataframe filled with binary values
}
\description{
Filling the dataframe with binary values. 0 for non differentially expressed genes and 1 for the other.
}
\examples{
#' # Import a dataset (Microarrays for example)
Data = Simul.data(type = "Microarrays", 
                    n.cond1 = 15, 
                    n.cond2 = 15, 
                    nb.genes = 100)
          
# Compute Pvalues for all the methods 
res.DEG = tools.DEG.Microarrays.merge(Data)
# Make a binary matrix to construct an Upset plot
Upset = Upset.Binary.Dataframe(res.DEG)
upset(Upset, sets = names(Upset), 
         sets.bar.color = "#56B4E9", 
         order.by = "freq", 
         empty.intersections = NULL )
}
