% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Upset.Binary.Dataframe}
\alias{Upset.Binary.Dataframe}
\title{Compute a usable matrix to make an Upset plot}
\usage{
Upset.Binary.Dataframe(data.to.comp, threshold)
}
\arguments{
\item{data.to.comp}{Dataframe of DEG pvalues with genes in columns, and methods in rows.}

\item{threshold}{Integer. 
For the whole dataframe, if value <= threshold, value = 1, else value = 0.}
}
\value{
dataframe filled with binary values
}
\description{
Fill the dataframe with binary values. 0 for non differentially expressed genes and 1 for the other.
}
\examples{
# Import the dataset
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 

# Compute Pvalues for all the methods 
# res.DEG = tools.DEG.RNAseq.merge(Data)
# Make a binary matrix to construct an Upset plot
# Upset = Upset.Binary.Dataframe(res.DEG)
#upset(Upset, sets = names(Upset), 
#          sets.bar.color = "#56B4E9", 
#          order.by = "freq", 
#          empty.intersections = NULL )
}
