% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{wilcoxDEG}
\alias{wilcoxDEG}
\title{Wilcoxon test for each row of a dataframe}
\usage{
wilcoxDEG(data, n1, n2)
}
\arguments{
\item{data}{dataframe of gene expression levels: Gene names in rows, samples in columns.}

\item{n1}{Number of samples for the first experimental condition}

\item{n2}{Number of samples for the second experimental condition}
}
\value{
dataframe with pvalues in one column and rownames of data as rownames.
}
\description{
Testing, for one row at the time, if the first series of values are different, greater or less than the values 
of the second condition with the \link[stats]{wilcox.test}
}
\examples{
# Import the dataset
Data = matrix(runif(5000, 10, 100), ncol=20)
group = paste0(rep(c("control", "case"), each = 10),rep(c(1:10),each = 1))
genes <- paste0(rep(LETTERS[1:25], each=10), rep(c(1:10),each = 1))
colnames(Data) = group
row.names(Data) = genes 

# Compute Pvalues
res.DEG = wilcoxDEG(Data,10,10)
}
